Build the cover code
Two compile errors, both from writing it without a compiler to hand.

FVector::Dist2D returns FVector::FReal, which is double in UE5, so putting it
straight into a float field of a braced initialiser is a narrowing conversion.

AActor already has a member called Owner, and a local of that name inside the
claim loops shadows it - a warning everywhere else, an error here because the
project builds warnings-as-errors.

Result: Succeeded, 11.5s.

+++ Polarity/AI/Components/CoverFinderComponent.cpp
@@ -141,7 +141,9 @@ void UCoverFinderComponent::OnQueryFinished(TSharedPtr<FEnvQueryResult> Result)
 			continue;
 		}
 
-		Scored.Add({ Candidate, ComputeExposure(Candidate, Players, TraceParams), FVector::Dist2D(Candidate, OwnerLocation) });
+		Scored.Add({ Candidate,
+			ComputeExposure(Candidate, Players, TraceParams),
+			static_cast<float>(FVector::Dist2D(Candidate, OwnerLocation)) });
 	}
 
 	if (Scored.Num() == 0)
